# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Интернет-магазин API",
        "description": "Class Controller",
        "version": "0.1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "authLogin",
                "requestBody": {
                    "$ref": "App/ApiDoc/Auth/LoginRequest"
                },
                "responses": {
                    "200": {
                        "description": "Success auth response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authorized. Wrong email or password.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "authRegister",
                "requestBody": {
                    "$ref": "App/ApiDoc/Auth/RegisterRequest"
                },
                "responses": {
                    "200": {
                        "description": "User successfully registered"
                    },
                    "500": {
                        "description": "Server error"
                    },
                    "422": {
                        "description": "Validation errors.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LoginSuccessResponse": {
                "title": "Login response",
                "description": "Success auth response",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "message": {
                        "type": "string",
                        "example": "User signed in."
                    },
                    "data": {
                        "title": "Validation errors",
                        "description": "Validation errors object",
                        "properties": {
                            "user_name": {
                                "type": "string"
                            },
                            "token": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UnauthenticatedResponse": {
                "title": "Unauthenticated response",
                "description": "Unauthenticated response schema",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ValidationResponse": {
                "title": "Validation errors response",
                "description": "Validation errors response",
                "properties": {
                    "message": {
                        "title": "Message",
                        "description": "Validation message",
                        "type": "string",
                        "format": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "title": "Validation errors",
                        "description": "Validation errors object",
                        "properties": {
                            "field1": {
                                "type": "array",
                                "items": {
                                    "example": "The field1 field is required."
                                }
                            },
                            "field2": {
                                "type": "array",
                                "items": {
                                    "example": "The field2 field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Class AuthController"
        }
    ]
}